reorganized the gui package
