+++ src/main/java/nodemanager/gui/editPage/mapComponents/MapImage.java
@@ -1,4 +1,4 @@
-package nodemanager.gui.mapComponents;
+package nodemanager.gui.editPage.mapComponents;
 
 import java.awt.*;
 import javax.swing.*;
@@ -8,7 +8,6 @@ import java.util.*;
 import nodemanager.model.Node;
 import nodemanager.*;
 import nodemanager.events.*;
-import nodemanager.gui.Scale;
 import nodemanager.model.Graph;
 
 /**
